requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
deep-translator==1.11.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
Flask==3.0.0
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
plotly==5.18.0
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
WTForms==3.1.2
